Fix some warnings and add clarification comment for ignore rule

+++ AppCenterCrashes/AppCenterCrashesTests/MSWrapperExceptionManagerTests.m
@@ -98,7 +98,8 @@ static NSString *const kMSLastWrapperExceptionFileName =
         .andReturn([[NSUUID UUID] UUIDString]);
     [mockReports addObject:reportMock];
   }
-  MSErrorReport *report = [mockReports objectAtIndex:(rand() % numReports)];
+  MSErrorReport *report =
+      [mockReports objectAtIndex:(NSUInteger)(rand() % numReports)];
   MSWrapperException *wrapperException = [self getWrapperException];
   wrapperException.processId =
       [NSNumber numberWithUnsignedInteger:[report appProcessIdentifier]];
